Renames "matchRequest" utility into "matchRequestUrl"
